% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiIndicators.R
\name{apply_delay}
\alias{apply_delay}
\title{\code{apply_delay}}
\usage{
apply_delay(g, s)
}
\arguments{
\item{g}{indicator.}

\item{s}{delay to apply.}
}
\value{
{ 
  A numeric vector with the result of apply the cevtor delay s to g. 
}
}
\description{
apply a delay vector, s(t), to an indicator, g(t)
}
