% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiIndicators.R
\name{EpiIndicators_params}
\alias{EpiIndicators_params}
\title{\code{EpiIndicators_params} function to select EpiIndicators parameters}
\usage{
EpiIndicators_params(x = "")
}
\arguments{
\item{x}{a list with parameters of EpiIndicators() function}
}
\value{
the EpiIndicators() parameters
}
\description{
\code{EpiIndicators_params} function to select EpiIndicators parameters
}
\seealso{
\code{\link{EpiIndicators}}
}
