% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{EpiIndicators_plot}
\alias{EpiIndicators_plot}
\title{\code{EpiIndicators_plot}}
\usage{
EpiIndicators_plot(df)
}
\arguments{
\item{df}{: a dataframe obtained as the outcome of executing EpiIndicators()}
}
\value{
a combination of 3 plots: (A) the indicator f(t) (in green in the main
y-axis) and the indicator g(t) (in red in the secondary axis). (B) r(t)*f(t) 
(in green in the main y-axis) and g(t+s(t)) (in red in the secondary axis) and  
(C) r(t) (in green in the main y-axis) and s(t) (in red in the secondary axis)
}
\description{
EpiIndicators_plot() plots the results obtained by EpiIndicators()
}
