% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{festives}
\alias{festives}
\title{A dataset containing festive days in  France, 
Germany, UK and the USA}
\format{
A list with 4 variables:
\describe{
  \item{FRA}{festive day of France}
  \item{DEU}{festive day  of Germany}
  \item{UK}{festive day  of UK}
  \item{USA}{festive day  of USA}
 
}
}
\usage{
festives
}
\description{
A dataset containing festive days in  France, 
Germany, UK and the USA
}
\keyword{datasets}
