% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiIndicators.R
\name{joint_indicators_by_date}
\alias{joint_indicators_by_date}
\title{\code{joint_indicators_by_date}}
\usage{
joint_indicators_by_date(date0, i0, date1, i1)
}
\arguments{
\item{date0}{the dates of the first indicator.}

\item{i0}{the values of the first indicator.}

\item{date1}{the dates of the second indicator.}

\item{i1}{the values of the second indicator.}
}
\value{
{ 
  A dataframe with the following columns :
  \itemize{

  \item{date}{: all dates presented in any of the indicators. }

  \item{f}{: the values of the first indicator. We assign 0
  in the case the data is not available for a given day.}
  
  \item{g}{: the values of the second indicator. We assign 0
  in the case the data is not available for a given day }

  }
}
}
\description{
generates a dataframe joining the dates and values
of 2 indicators
}
