% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owid}
\alias{owid}
\title{A dataset with COVID-19 indicators}
\format{
A dataframe with 13 variables:
\describe{
  \item{iso_code}{iso code of the country}
  \item{location}{country name}
  \item{date}{date of the indicator value }
  \item{new_cases}{new confirmed cases}
  \item{new_cases_smoothed}{new confirmed cases smoothed}
  \item{new_cases_restored_EpiInvert}{new confirmed cases restored using EpiInvert}
  \item{new_deaths}{new deaths attributed to COVID-19}
  \item{new_deaths_smoothed}{new deaths smoothed}
  \item{new_deaths_restored_EpiInvert}{new deaths restored using EpiInvert}
  \item{icu_patients}{number of COVID-19 patients in intensive care units (ICUs) 
  on a given day}
  \item{hosp_patients}{number of COVID-19 patients in hospital on a given day}
  \item{weekly_icu_admissions}{number of COVID-19 patients newly admitted to 
  intensive care units (ICUs) in a given week (reporting date and the preceding 
  6 days)}
  \item{weekly_hosp_admissions}{number of COVID-19 patients newly admitted to 
hospitals in a given week (reporting date and the preceding 6 days)}
 
}
}
\usage{
owid
}
\description{
A dataset containing COVID-19 epidemiological indicators for Canada, France, 
Germany, Italy, UK and the USA from Our World in data organization
https://github.com/owid/covid-19-data/tree/master/public/data up to 2022-11-28. 
In the case a data value is not available for a given day we assign the value 0 
to the indicator
}
\keyword{datasets}
