% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{\code{plot} the results obtained by EpiInvert}
\usage{
plot(x, what = "all", date_start = "1000-01-01", date_end = "3000-01-01")
}
\arguments{
\item{x}{an object of class \code{estimate_EpiInvert}.}

\item{what}{one of the following drawing options: 

 \itemize{
 \item{all}{: a plot combining the main EpiInvert results}
 \item{R}{: a plot of the reproduction number Rt estimation }
 \item{incid}{: a plot combining the obtained incidence curves }
 \item{SI}{: the serial interval used in the EpiInvert estimation }
 
}}

\item{date_start}{the start date to plot}

\item{date_end}{the final date to plot}
}
\value{
a plot.
}
\description{
\code{plot} the results obtained by EpiInvert
}
\seealso{
\code{\link{EpiInvert}}
}
