% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{restored_incidence_database}
\alias{restored_incidence_database}
\title{A dataset including 27,418 samples of different
restored incidence curves computed by EpiInvert using real data. Each 
restored incidence curve includes the last 56 values of the sequence.}
\format{
A 27,418 X 56 numeric matrix
}
\usage{
restored_incidence_database
}
\description{
A dataset including 27,418 samples of different
restored incidence curves computed by EpiInvert using real data. Each 
restored incidence curve includes the last 56 values of the sequence.
}
\keyword{datasets}
