% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_params.R
\name{select_params}
\alias{select_params}
\title{\code{select_params} function to select EpiInvert parameters}
\usage{
select_params(x = "")
}
\arguments{
\item{x}{a list with elements of the class  \code{estimate_R_config}}
}
\value{
an object of class \code{estimate_R_config}.
}
\description{
\code{select_params} function to select EpiInvert parameters
}
\seealso{
\code{\link{EpiInvert}}
}
