% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_distr_data}
\alias{si_distr_data}
\title{A dataset containing the values of a serial interval}
\format{
A numeric vector with 1 variable representing 
the serial interval
}
\usage{
si_distr_data
}
\description{
A dataset containing the values of a serial interval
}
\keyword{datasets}
