% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{EpiInvertForecast_plot}
\alias{EpiInvertForecast_plot}
\title{\code{EpiInvertForecast_plot}}
\usage{
EpiInvertForecast_plot(EpiInvert_results, Forecast)
}
\arguments{
\item{EpiInvert_results}{the list returned by the EpiInvert execution}

\item{Forecast}{the list returned by the EpiInvertForecast execution}
}
\value{
a plot with the last 28 days of the original and restored incidence curves and a 28-day 
forecast of the same curves. It also includes a shaded area with a 95% empiric confidence interval
of the restored incidence forecast estimation
}
\description{
EpiInvertForecast_plot() plot the restored incidence forecast
}
